extracted tasks into separate classes
