+++ gradle-mpsbuild-plugin/src/main/kotlin/org/modelix/gradle/mpsbuild/MPSBuildPlugin.kt
@@ -13,297 +13,100 @@
  */
 package org.modelix.gradle.mpsbuild
 
-import org.gradle.api.Action
 import org.gradle.api.Plugin
 import org.gradle.api.Project
-import org.gradle.api.Task
-import org.gradle.api.artifacts.Configuration
-import org.gradle.api.artifacts.ResolvedConfiguration
 import org.gradle.api.artifacts.ResolvedDependency
 import org.gradle.api.publish.PublishingExtension
 import org.gradle.api.publish.maven.MavenPublication
 import org.gradle.api.publish.maven.tasks.GenerateMavenPom
 import org.gradle.api.publish.maven.tasks.PublishToMavenRepository
 import org.gradle.api.tasks.Exec
+import org.gradle.api.tasks.TaskProvider
 import org.modelix.buildtools.*
-import org.zeroturnaround.zip.ZipUtil
 import java.io.File
 import java.nio.file.Path
 import java.nio.file.Paths
 import java.text.SimpleDateFormat
 import java.util.*
-import java.util.stream.Collectors
-import java.util.zip.ZipEntry
-import java.util.zip.ZipOutputStream
-import kotlin.text.uppercase
 
 class MPSBuildPlugin : Plugin<Project> {
-    private val stubsPattern = Regex("stubs#([^#]+)#([^#]+)#([^#]+)")
+
     private lateinit var project: Project
     private lateinit var settings: MPSBuildSettings
-    private val folder2owningDependency = HashMap<Path, ResolvedDependency>()
-
-    lateinit var taskCopyDependencies: Task
-    lateinit var taskGenerateAntScript: Task
-    lateinit var taskCheckConfig: Task
-    lateinit var taskLoadPomDependencies: Task
-    lateinit var taskPackagePublications: Task
-
-    private fun newTask(name: String): Task = project.task(name)
-
-    private fun newTask(name: String, body: ()->Unit): Task {
-        return project.task(name) {
-            val action = Action<Task> { body() }
-            this.actions = listOf(action)
-        }
-    }
-
-    private fun taskBody(task: Task, body: ()->Unit) {
-        val action = Action<Task> { body() }
-        task.actions = listOf(action)
-    }
+    private lateinit var buildDir: File
+    private lateinit var publicationsDir: File
+    private lateinit var dependenciesDir: File
+    private lateinit var antScriptFile: File
 
     override fun apply(project: Project) {
         this.project = project
         settings = project.extensions.create("mpsBuild", MPSBuildSettings::class.java)
         settings.setProject(project)
+        buildDir = project.buildDir.resolve("mpsbuild").normalize()
+        dependenciesDir = buildDir.resolve("dependencies")
+        publicationsDir = buildDir.resolve("publications")
+        antScriptFile = File(buildDir, "build-modules.xml")
 
-        taskCopyDependencies = newTask("copyDependencies")
-        taskGenerateAntScript = newTask("generateMpsAntScript")
-        taskCheckConfig = newTask("checkMpsbuildConfig")
-        taskLoadPomDependencies = newTask("loadPomDependencies")
-        taskPackagePublications = newTask("packageMpsPublications")
-
-        project.afterEvaluate {
-            settings.validate()
-            this@MPSBuildPlugin.afterEvaluate()
+        val folder2owningDependency = HashMap<Path, ResolvedDependency>()
+        val taskCopyDependencies = project.tasks.register("copyDependencies", CopyDependencies::class.java) {
+            settings.getTaskDependencies().forEach { this.dependsOn(it) }
+            this.dependenciesConfig.set(settings.dependenciesConfig)
+            this.dependenciesTargetDir.set(dependenciesDir.normalize())
+            this.folderToOwningDependency.set(folder2owningDependency)
+            this.mpsDir.set(buildDir.resolve("mps"))
         }
-    }
+        val dirsToMine = setOfNotNull(dependenciesDir, taskCopyDependencies.get().mpsDir.asFile.get())
+        val buildScriptGenerator = createBuildScriptGenerator(settings, project, buildDir, dirsToMine)
 
-    private fun afterEvaluate() {
-        val buildDir = project.buildDir.resolve("mpsbuild").normalize()
-        val dependenciesDir = buildDir.resolve("dependencies")
-        val publicationsDir = buildDir.resolve("publications")
-        val antScriptFile = File(buildDir, "build-modules.xml")
-        val publicationsVersion = getPublicationsVersion()
-        val mavenPublications = HashMap<MPSBuildSettings.PublicationSettings, MavenPublication>()
-
-        var mpsDir: File? = null
-
-        taskBody(taskCopyDependencies) {
-            copyDependencies(settings.dependenciesConfig, dependenciesDir.normalize())
-            mpsDir = downloadMps(settings, buildDir.resolve("mps"))
+        val taskGenerateAntScript = project.tasks.register("generateMpsAntScript", GenerateAntScript::class.java) {
+            this.dependsOn(taskCopyDependencies)
+            this.generator.set(buildScriptGenerator)
+            this.antFile.set(antScriptFile)
         }
-        settings.getTaskDependencies().forEach { taskCopyDependencies.dependsOn(it) }
-
-        lateinit var generator: BuildScriptGenerator
-        taskBody(taskGenerateAntScript) {
-            val dirsToMine = setOfNotNull(dependenciesDir, mpsDir)
-            generator = createBuildScriptGenerator(settings, project, buildDir, dirsToMine)
-            generateAntScript(generator, antScriptFile)
+        val taskCheckConfig = project.tasks.register("checkMpsbuildConfig", CheckConfig::class.java) {
+            this.generator.set(buildScriptGenerator)
+            this.settings.set(this@MPSBuildPlugin.settings)
         }
-        taskGenerateAntScript.dependsOn(taskCopyDependencies)
-
-        lateinit var publication2dnode: Map<MPSBuildSettings.PublicationSettings, DependencyGraph<FoundModule, ModuleId>.DependencyNode>
-        lateinit var getPublication: (DependencyGraph<FoundModule, ModuleId>.DependencyNode)->MPSBuildSettings.PublicationSettings?
-
-        taskBody(taskCheckConfig) {
-            val resolver = ModuleResolver(generator.modulesMiner.getModules(), generator.ignoredModules)
-            val graph = PublicationDependencyGraph(resolver, emptyMap())
-            val publication2modules = settings.getPublications().associateWith { resolvePublicationModules(it, resolver).toSet() }
-            for (modulesA in publication2modules) {
-                for (modulesB in publication2modules) {
-                    if (modulesA.key == modulesB.key) continue
-                    val modulesInBoth = modulesA.value.intersect(modulesB.value)
-                    require(modulesInBoth.isEmpty()) {
-                        "Modules found in publication ${modulesA.key.name} and ${modulesB.key.name}: ${modulesInBoth.map { it.name }.sorted()}"
-                    }
-                }
-            }
-            graph.load(publication2modules.values.flatten())
-            val module2publication = publication2modules.flatMap { entry -> entry.value.map { it to entry.key } }.associate { it }
-
-            getPublication = {
-                it.modules.mapNotNull { module2publication[it] }.firstOrNull()
-            }
-
-            val checkCyclesBetweenPublications = {
-                val cycleDetection = object : CycleDetection<DependencyGraph<FoundModule, ModuleId>.DependencyNode, MPSBuildSettings.PublicationSettings>() {
-                    override fun getOutgoingEdges(element: DependencyGraph<FoundModule, ModuleId>.DependencyNode): Iterable<DependencyGraph<FoundModule, ModuleId>.DependencyNode> {
-                        return element.getDependencies()
-                    }
-
-                    override fun getCategory(element: DependencyGraph<FoundModule, ModuleId>.DependencyNode): MPSBuildSettings.PublicationSettings? {
-                        return getPublication(element)
-                    }
-                }
-                cycleDetection.process(graph.getNodes())
-                for (cycle in cycleDetection.cycles) {
-                    val pubs = cycle.mapNotNull { getPublication(it) }.distinct()
-                    if (pubs.size > 1) {
-                        throw RuntimeException("Cycle between publications ${pubs.joinToString(" and ") { it.name } } probably caused by these modules: " + cycle.map { it.modules.map { it.name } })
-                    }
-                }
-            }
-            checkCyclesBetweenPublications()
-            publication2dnode = publication2modules.entries.associate {
-                it.key to graph.mergeElements(it.value)
-            }
-            checkCyclesBetweenPublications()
-
-            val ensurePublicationsNotMerged: ()->Unit = {
-                for (publicationA in publication2dnode) {
-                    for (publicationB in publication2dnode) {
-                        if (publicationA.key == publicationB.key) continue
-                        require(publicationA.value.getMergedNode() != publicationB.value.getMergedNode()) {
-                            "Unexpected merge of publications '${publicationA.key.name}' and '${publicationB.key.name}'"
-                        }
-                    }
-                }
-            }
-
-            ensurePublicationsNotMerged()
-            graph.mergeCycles()
-            ensurePublicationsNotMerged()
-
-            // merge nodes with exclusive direct dependency between them
-            while (true) {
-                var anyMerge = false
-                for (n in graph.getNodes().filter { it.getReverseDependencies().size == 1 }) {
-                    if (n.modules.all { it.owner !is SourceModuleOwner }) continue
-                    val reverseDependencies = n.getReverseDependencies()
-                    if (reverseDependencies.size != 1) continue // may have changed, because this loop modifies the graph
-                    if (publication2dnode.values.map { it.getMergedNode() }.contains(n)) continue
-                    graph.mergeNodes(n, reverseDependencies.first())
-                    anyMerge = true
-                }
-                if (!anyMerge) break
-            }
-
-            ensurePublicationsNotMerged()
-
-            for (node in graph.getNodes()) {
-                val modules = node.modules
-                    .filter { it.owner is SourceModuleOwner }
-                    .map { it.name }
-                    .filter { !it.startsWith("stubs#") }
-                    .sorted()
-                if (modules.isEmpty()) continue
-                val publication = getPublication(node)
-                require(publication != null) {
-                    "Module $modules is used by multiple publications ${node.getReverseDependencies().mapNotNull(getPublication).map { it.name }}, but not part of any publication itself."
-                }
-//                println("Publication ${publication.name}")
-//                for (module in modules) {
-//                    println("    $module")
-//                }
-            }
+        val taskLoadPomDependencies = project.tasks.register("loadPomDependencies", LoadPomDependencies::class.java) {
+            this.dependsOn(taskCheckConfig)
+            this.settings.set(this@MPSBuildPlugin.settings)
+            this.publicationToDnode.set(taskCheckConfig.get().publication2dnode)
+            this.folderToOwningDependency.set(folder2owningDependency)
+            this.publicationsVersion.set(getPublicationsVersion())
+            this.getPublication.set(taskCheckConfig.get().getPublication)
         }
-        taskCheckConfig.dependsOn(taskGenerateAntScript)
-
-        val taskAssembleMpsModules = project.tasks.create("assembleMpsModules", Exec::class.java) {
+        val taskAssembleMpsModules = project.tasks.register("assembleMpsModules", Exec::class.java) {
+            this.dependsOn(taskGenerateAntScript)
+            this.mustRunAfter(taskCheckConfig) // fail fast
             workingDir = antScriptFile.parentFile
             commandLine = listOf("ant", "-f", antScriptFile.absolutePath, "assemble")
             standardOutput = System.out
             errorOutput = System.err
             standardInput = System.`in`
         }
-        taskAssembleMpsModules.dependsOn(taskGenerateAntScript)
 
-        taskBody(taskLoadPomDependencies) {
-            for (publication in settings.getPublications()) {
-                val dnode = publication2dnode[publication]!!.getMergedNode()
-                val pluginModuleNames = settings.getPluginModuleNames()
-                val modulesAndStubs = dnode.modules.filter { !pluginModuleNames.contains(it.name) }
-                val stubs = modulesAndStubs.filter { it.name.startsWith("stubs#") }.toSet()
-                mavenPublications[publication]?.pom {
-                    withXml {
-                        asElement().newChild("dependencies") {
-                            // dependencies between own publications
-                            for (dependency in dnode.getDependencies().mapNotNull(getPublication)) {
-                                newChild("dependency") {
-                                    newChild("groupId", project.group.toString())
-                                    newChild("artifactId", dependency.name.toValidPublicationName())
-                                    newChild("version", publicationsVersion)
-                                }
-                            }
-
-                            // dependencies to downloaded publications
-                            val externalDependencies = (dnode.getDependencies() + dnode).flatMap { it.modules }
-                                .mapNotNull { getOwningDependency(it.owner.path.getLocalAbsolutePath()) }
-                                .distinct()
-                            for (dependency in externalDependencies) {
-                                newChild("dependency") {
-                                    newChild("groupId", dependency.moduleGroup)
-                                    newChild("artifactId", dependency.moduleName)
-                                    newChild("version", dependency.moduleVersion)
-                                }
-                            }
-
-                            // dependencies to java libraries
-                            for (stub in stubs) {
-                                val match = stubsPattern.matchEntire(stub.name)
-                                    ?: throw RuntimeException("Failed to extract maven coordinates from ${stub.name}")
-                                newChild("dependency") {
-                                    newChild("groupId", match.groupValues[1])
-                                    newChild("artifactId", match.groupValues[2])
-                                    newChild("version", match.groupValues[3])
-                                }
-                            }
-                        }
-                    }
-                }
-            }
+        val taskPackagePublications = project.tasks.register("packageMpsPublications", PackageMpsPublications::class.java) {
+            this.dependsOn(taskCheckConfig)
+            this.dependsOn(taskLoadPomDependencies)
+            this.dependsOn(taskAssembleMpsModules)
+            this.publicationsDir.set(this@MPSBuildPlugin.publicationsDir)
+            this.generator.set(buildScriptGenerator)
+            this.settings.set(this@MPSBuildPlugin.settings)
+            this.publicationsVersion.set(getPublicationsVersion())
+            this.publication2dnode.set(taskLoadPomDependencies.get().publicationToDnode)
         }
-        taskLoadPomDependencies.dependsOn(taskCheckConfig)
-
-        taskBody(taskPackagePublications) {
-            val packagedModulesDir = generator.getPackagedModulesDir()
-            val generatedPlugins = generator.getGeneratedPlugins().entries.associate { it.key.name to it.value }
-            val pluginModuleNames = settings.getPluginModuleNames()
-            for (publication in settings.getPublications()) {
-                val dnode = publication2dnode[publication]!!.getMergedNode()
-                val modulesAndStubs = dnode.modules.filter { !pluginModuleNames.contains(it.name) }
-                val stubs = modulesAndStubs.filter { it.name.startsWith("stubs#") }.toSet()
-                val modules = modulesAndStubs - stubs
-                val generatedFiles = modules.map { it.owner }.filterIsInstance<SourceModuleOwner>()
-                    .distinct().flatMap { generator.getGeneratedFiles(it) }.map { it.absoluteFile.normalize() }
-                val zipFile = publicationsDir.resolve("${publication.name}.zip")
-                zipFile.parentFile.mkdirs()
-                zipFile.outputStream().use { os ->
-                    ZipOutputStream(os).use { zipStream ->
-                        val packFile: (File, Path, Path)->Unit = { file, path, parent ->
-                            val relativePath = parent.relativize(path).toString()
-                            require(!path.toString().startsWith("..") && !path.toString().contains("/../")) {
-                                "$file expected to be inside $parent"
-                            }
-                            val entry = ZipEntry(relativePath)
-                            zipStream.putNextEntry(entry)
-                            file.inputStream().use { istream -> istream.copyTo(zipStream) }
-                            zipStream.closeEntry()
-                        }
-                        for (file in generatedFiles) {
-                            packFile(file, file.toPath(), packagedModulesDir.parentFile.toPath())
-                        }
-                        for (ideaPlugin in publication.ideaPlugins) {
-                            val pluginFolder = generatedPlugins[ideaPlugin.getImplementationModuleName()]
-                                ?: throw RuntimeException("Output for plugin '${ideaPlugin.getImplementationModuleName()}' not found")
-                            for (file in pluginFolder.walk()) {
-                                if (file.isFile) {
-                                    packFile(file, file.toPath(), pluginFolder.parentFile.parentFile.toPath())
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-
-            println("Version $publicationsVersion ready for publishing")
+        project.tasks.withType(GenerateMavenPom::class.java).matching { it.name.matches(Regex(".+_.+_.+")) }.all {
+            dependsOn(taskLoadPomDependencies)
         }
-        taskPackagePublications.dependsOn(taskCheckConfig)
-        taskPackagePublications.dependsOn(taskLoadPomDependencies)
-        taskPackagePublications.dependsOn(taskAssembleMpsModules)
-        taskAssembleMpsModules.mustRunAfter(taskCheckConfig) // fail fast
+        project.afterEvaluate {
+            settings.validate()
+            this@MPSBuildPlugin.afterEvaluate(taskPackagePublications)
+        }
+    }
+
+    private fun afterEvaluate(taskPackagePublications: TaskProvider<PackageMpsPublications>) {
+        val publicationsVersion = getPublicationsVersion()
+        val mavenPublications = HashMap<MPSBuildSettings.PublicationSettings, MavenPublication>()
 
         val mpsPublicationsConfig = project.configurations.create("mpsPublications")
         val publishing = project.extensions.findByType(PublishingExtension::class.java)
@@ -346,10 +149,6 @@ class MPSBuildPlugin : Plugin<Project> {
             }
         }
 
-        project.tasks.withType(GenerateMavenPom::class.java).matching { it.name.matches(Regex(".+_.+_.+")) }.all {
-            dependsOn(taskLoadPomDependencies)
-        }
-
         val repositories = publishing?.repositories ?: listOf()
         val ownArtifactNames = settings.getPublications().map { it.name.toValidPublicationName() }.toSet() + "all"
         for (repo in repositories) {
@@ -376,41 +175,6 @@ class MPSBuildPlugin : Plugin<Project> {
         ("" + project.version).ifEmpty { null }
     } ?: generateVersionNumber(settings.mpsMajorVersion)
 
-    private fun downloadMps(settings: MPSBuildSettings, targetDir: File): File? {
-        var mpsDir: File? = null
-        settings.mpsDependenciesConfig?.resolvedConfiguration?.lenientConfiguration?.let {
-            for (file in it.files) {
-                val targetFile = targetDir.resolve(file.name)
-                if (!targetFile.exists()) {
-                    copyAndUnzip(file, targetFile)
-                }
-                mpsDir = targetFile
-            }
-        }
-
-        if (mpsDir == null) {
-            val url = settings.getMpsDownloadUrl()
-            if (url != null) {
-                val file = targetDir.resolve(url.toString().substringAfterLast("/"))
-                if (!file.exists()) {
-                    println("Downloading $url")
-                    file.parentFile.mkdirs()
-                    url.openStream().use { istream ->
-                        file.outputStream().use { ostream ->
-                            istream.copyTo(ostream)
-                        }
-                    }
-                }
-                if (file.isFile) {
-                    ZipUtil.explode(file)
-                }
-                mpsDir = file
-            }
-        }
-
-        return mpsDir
-    }
-
     private fun generateVersionNumber(mpsVersion: String?): String {
         val timestamp = SimpleDateFormat("yyyyMMddHHmm").format(Date())
         val version = if (mpsVersion == null) timestamp else "$mpsVersion-$timestamp"
@@ -418,83 +182,7 @@ class MPSBuildPlugin : Plugin<Project> {
         return version
     }
 
-    private fun generateStubsSolution(dependency: ResolvedDependency, stubsDir: File) {
-        val solutionName = getStubSolutionName(dependency)
-        val jars = dependency.moduleArtifacts.map { it.file }.filter { it.extension == "jar" }
-        val xml = newXmlDocument {
-            newChild("solution") {
-                setAttribute("name", solutionName)
-                setAttribute("pluginKind", "PLUGIN_OTHER")
-                setAttribute("moduleVersion", "0")
-                setAttribute("uuid", "~$solutionName")
-                newChild("facets") {
-                    newChild("facet") {
-                        setAttribute("type", "java")
-                    }
-                }
-                newChild("models") {
-                    for (jar in jars) {
-                        newChild("modelRoot") {
-                            setAttribute("type", "java_classes")
-                            setAttribute("contentPath", jar.parentFile.absolutePath)
-                            newChild("sourceRoot") {
-                                setAttribute("location", jar.name)
-                            }
-                        }
-                    }
-                }
-                newChild("dependencies") {
-                    newChild("dependency", "6354ebe7-c22a-4a0f-ac54-50b52ab9b065(JDK)") {
-                        setAttribute("reexport", "true")
-                    }
-                    for (transitiveDep in dependency.children) {
-                        val n = getStubSolutionName(transitiveDep)
-                        newChild("dependency", "~$n($n)") {
-                            setAttribute("reexport", "true")
-                        }
-                    }
-                }
-                newChild("stubModelEntries") {
-                    for (jar in jars) {
-                        newChild("stubModelEntry") {
-                            setAttribute("path", jar.absolutePath)
-                        }
-                    }
-                }
-            }
-        }
-        val solutionFile = stubsDir.resolve(solutionName).resolve("$solutionName.msd")
-        solutionFile.parentFile.mkdirs()
-        solutionFile.writeText(xmlToString(xml))
-    }
-
     private fun String.toValidPublicationName() = replace(Regex("[^A-Za-z0-9_\\-.]"), "_").toLowerCase()
-
-    private fun getStubSolutionName(dependency: ResolvedDependency): String {
-//                        val clean: (String)->String = { it.replace(Regex("[^a-zA-Z0-9]"), "_") }
-//                        val group = clean(it.moduleGroup)
-//                        val artifactId = clean(it.moduleName)
-//                        val version = clean(it.moduleVersion)
-//                        "stubs.$group.$artifactId.$version"
-        return "stubs#" + dependency.module.id.toString().replace(":", "#")
-    }
-
-    private fun ResolvedConfiguration.getAllDependencies(): List<ResolvedDependency> {
-        val allDependencies: MutableList<ResolvedDependency> = ArrayList()
-        object : GraphWithCyclesVisitor<ResolvedDependency>() {
-            override fun onVisit(element: ResolvedDependency) {
-                allDependencies.add(element)
-                visit(element.children)
-            }
-        }.visit(firstLevelModuleDependencies)
-        return allDependencies
-    }
-
-    private fun generateAntScript(generator: BuildScriptGenerator, antScriptFile: File): BuildScriptGenerator {
-        val xml = generator.generateXML()
-        antScriptFile.writeText(xml)
-        return generator
-    }
 
     private fun createBuildScriptGenerator(settings: MPSBuildSettings,
                                            project: Project,
@@ -561,54 +249,6 @@ class MPSBuildPlugin : Plugin<Project> {
             throw RuntimeException("Modules not found: $missingModuleNames")
         }
         return modulesToGenerate
-    }
-
-    private fun copyDependencies(dependenciesConfiguration: Configuration, targetFolder: File) {
-        val dependencies = dependenciesConfiguration.resolvedConfiguration.getAllDependencies()
-        for (dependency in dependencies) {
-            val files = dependency.moduleArtifacts.map { it.file }
-            for (file in files) {
-                copyDependency(file, targetFolder, dependency)
-            }
-        }
-    }
-
-    private fun copyDependency(file: File, targetFolder: File, dependency: ResolvedDependency) {
-        when (file.extension) {
-            "jar" -> {
-                generateStubsSolution(dependency, targetFolder.resolve("stubs"))
-            }
-            "zip" -> {
-                val targetFile = targetFolder
-                    .resolve("modules")
-                    .resolve(dependency.moduleGroup)
-                    .resolve(dependency.moduleName)
-                    //.resolve(file.name)
-                folder2owningDependency[targetFile.absoluteFile.toPath().normalize()] = dependency
-                copyAndUnzip(file, targetFile)
-            }
-            else -> println("Ignored file $file from dependency ${dependency.module.id}")
-        }
-    }
-
-    private fun getOwningDependency(file: Path): ResolvedDependency? {
-        var f: Path? = file.toAbsolutePath().normalize()
-        while (f != null) {
-            val owner = folder2owningDependency[f]
-            if (owner != null) return owner
-            f = f.parent
-        }
-        return null
-    }
-
-    private fun copyAndUnzip(sourceFile: File, targetFile: File) {
-        targetFile.parentFile.mkdirs()
-        if (sourceFile.extension == "zip") {
-            if (targetFile.exists()) targetFile.deleteRecursively()
-            ZipUtil.unpack(sourceFile, targetFile)
-        } else {
-            sourceFile.copyTo(targetFile, true)
-        }
     }
 }
